// src/types/biodiversity.ts
export type MigrationType = 'breeder' | 'winter-visitor' | 'summer-visitor' | 'passage-migrant';

export type Flyway = 'central-asian' | 'east-asian' | 'west-asian';